Add update products
Reformated code with black. Also renamed ds_app to api.
