+++ api/dao/customer_dao.py
@@ -1,9 +1,9 @@
 import psycopg2
 
+
 class CustomerDAO(object):
     def __init__(self):
         return
-    
+
     def insert_customer(self):
         return
-
