+++ api/dao/location_dao.py
@@ -1,0 +1,13 @@
+import psycopg2
+
+
+class LocationDAO(object):
+    def __init__(self):
+        return
+
+    def insert_location(self, latitude, longitude):
+        location_id = 5
+        return location_id
+
+    def update_location(self, location_id, latitude, longitude):
+        return location_id
