+++ api/handlers/customer_handler.py
@@ -1,12 +1,10 @@
-from ds_app.dao.customer_dao import CustomerDAO
+from api.dao.customer_dao import CustomerDAO
 from flask import jsonify
 
 
 class CustomerHandler(object):
-
     def build_customer_attributes(self, *attr):
         return
 
     def add_customer(self, payload):
         return
-
