--- ds_app/dao/location_dao.py
@@ -1,11 +1,0 @@
-import psycopg2
-
-class LocationDAO(object):
-
-    def __init__(self):
-        return
-
-
-    def insert_location(self, latitude, longitude):
-        location_id = 5
-        return  location_id
