+++ main.py
@@ -1,8 +1,8 @@
 from flask import Flask, jsonify, request
 from flask_cors import CORS, cross_origin
-from ds_app.handlers.product_handler import ProductHandler
-from ds_app.handlers.customer_handler import CustomerHandler
-from ds_app.handlers.user_handler import UserHandler
+from api.handlers.product_handler import ProductHandler
+from api.handlers.customer_handler import CustomerHandler
+from api.handlers.user_handler import UserHandler
 
 app = Flask(__name__)
 CORS(app)
@@ -15,22 +15,24 @@ def greet():
 
 ### USERS ###
 
-@app.route('/disasterStorage/users',  methods=['GET'])
+
+@app.route("/disasterStorage/users", methods=["GET"])
 def get_all_users():
-		if request.method == 'GET':
-			return UserHandler().get_all_users()
-		return
+    if request.method == "GET":
+        return UserHandler().get_all_users()
+    return
 
-@app.route('/disasterStorage/users/<int:user_id>', methods=['GET', 'PUT'])
+
+@app.route("/disasterStorage/users/<int:user_id>", methods=["GET", "PUT"])
 def get_user_by_id(user_id):
-	if request.method == 'GET':
-		return UserHandler().get_user_by_id(user_id)
+    if request.method == "GET":
+        return UserHandler().get_user_by_id(user_id)
 
-	elif request.method == 'PUT':
-		return
+    elif request.method == "PUT":
+        return
 
-	else:
-		return
+    else:
+        return
 
 
 ### CUSTOMERS ###
@@ -38,9 +40,10 @@ def get_user_by_id(user_id):
 
 @app.route("/disasterStorage/users/customers", methods=["GET", "POST"])
 def get_all_customers():
-	return
+    return
 
-@app.route('/disasterStorage/users/<int:customer_id>', methods=['GET', 'PUT', 'DELETE'])
+
+@app.route("/disasterStorage/users/<int:customer_id>", methods=["GET", "PUT", "DELETE"])
 def get_customer_by_id(customer_id):
     return
 
@@ -90,7 +93,9 @@ def get_all_products():
         return ProductHandler().insert_product(request.json)
 
 
-@app.route("/disasterStorage/products/<int:product_id>", methods=["GET", "PUT", "DELETE"])
+@app.route(
+    "/disasterStorage/products/<int:product_id>", methods=["GET", "PUT", "DELETE"]
+)
 def get_product_by_id(product_id):
     if request.method == "GET":
         if "d" in request.args:
@@ -99,10 +104,11 @@ def get_product_by_id(product_id):
             return ProductHandler().get_product_by_id(product_id)
 
     elif request.method == "PUT":
-        return
+        return ProductHandler().update_product(product_id, request.json)
 
     else:
         return
+
 
 @app.route("/disasterStorage/products/available")
 def get_available_products():
